requirements.txt
aiocqhttp
matplotlib
requests
requests_toolbelt
beautifulsoup4
feedparser
pandas
Pillow
